Fix config with not setting CONFIG_INTEL_ATOMISP:
It causes a kernel panic related to vlv2_plat_configure_clock.

Instead of probing whether the device is installed or not the driver's
probe function just assumes the device is present and maps a hard coded
physical address and begins working on that, regardless whether the
mapped address belongs to the correct device type or not.

Source: https://lists.xen.org/archives/html/xen-devel/2017-09/msg00793.html